Add an entrypoint script for the iceberg-maintenance command
